+++ src/main/java/com/pickyfy/pickyfy/service/EmailService.java
@@ -1,9 +1,9 @@
 package com.pickyfy.pickyfy.service;
 
-import com.pickyfy.pickyfy.dto.request.EmailVerificationSendRequest;
-import com.pickyfy.pickyfy.dto.request.EmailVerificationVerifyRequest;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationSendResponse;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationVerifyResponse;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationSendRequest;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationVerifyRequest;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationSendResponse;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationVerifyResponse;
 
 public interface EmailService {
