+++ src/main/java/com/pickyfy/pickyfy/service/EmailServiceImpl.java
@@ -3,10 +3,10 @@ package com.pickyfy.pickyfy.service;
 import com.pickyfy.pickyfy.apiPayload.code.status.ErrorStatus;
 import com.pickyfy.pickyfy.common.util.JwtUtil;
 import com.pickyfy.pickyfy.common.util.RedisUtil;
-import com.pickyfy.pickyfy.dto.request.EmailVerificationSendRequest;
-import com.pickyfy.pickyfy.dto.request.EmailVerificationVerifyRequest;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationSendResponse;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationVerifyResponse;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationSendRequest;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationVerifyRequest;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationSendResponse;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationVerifyResponse;
 import com.pickyfy.pickyfy.exception.handler.ExceptionHandler;
 import com.pickyfy.pickyfy.repository.UserRepository;
 import jakarta.mail.MessagingException;
