+++ src/main/java/com/pickyfy/pickyfy/service/UserService.java
@@ -1,7 +1,7 @@
 package com.pickyfy.pickyfy.service;
 
-import com.pickyfy.pickyfy.dto.request.UserCreateRequest;
-import com.pickyfy.pickyfy.dto.response.UserCreateResponse;
+import com.pickyfy.pickyfy.web.dto.request.UserCreateRequest;
+import com.pickyfy.pickyfy.web.dto.response.UserCreateResponse;
 
 public interface UserService {
