+++ src/main/java/com/pickyfy/pickyfy/web/controller/EmailControllerApi.java
@@ -1,10 +1,10 @@
-package com.pickyfy.pickyfy.controller;
+package com.pickyfy.pickyfy.web.controller;
 
 import com.pickyfy.pickyfy.apiPayload.ApiResponse;
-import com.pickyfy.pickyfy.dto.request.EmailVerificationSendRequest;
-import com.pickyfy.pickyfy.dto.request.EmailVerificationVerifyRequest;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationSendResponse;
-import com.pickyfy.pickyfy.dto.response.EmailVerificationVerifyResponse;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationSendRequest;
+import com.pickyfy.pickyfy.web.dto.request.EmailVerificationVerifyRequest;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationSendResponse;
+import com.pickyfy.pickyfy.web.dto.response.EmailVerificationVerifyResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.web.bind.annotation.RequestBody;
