+++ src/main/java/com/pickyfy/pickyfy/web/controller/UserControllerApi.java
@@ -1,8 +1,8 @@
-package com.pickyfy.pickyfy.controller;
+package com.pickyfy.pickyfy.web.controller;
 
 import com.pickyfy.pickyfy.apiPayload.ApiResponse;
-import com.pickyfy.pickyfy.dto.request.UserCreateRequest;
-import com.pickyfy.pickyfy.dto.response.UserCreateResponse;
+import com.pickyfy.pickyfy.web.dto.request.UserCreateRequest;
+import com.pickyfy.pickyfy.web.dto.response.UserCreateResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.web.bind.annotation.RequestBody;
