+++ src/main/java/com/pickyfy/pickyfy/web/dto/CustomUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.pickyfy.pickyfy.dto;
+package com.pickyfy.pickyfy.web.dto;
 
 import com.pickyfy.pickyfy.domain.User;
 import lombok.Builder;
