+++ src/main/java/com/pickyfy/pickyfy/web/dto/request/EmailVerificationVerifyRequest.java
@@ -1,4 +1,4 @@
-package com.pickyfy.pickyfy.dto.request;
+package com.pickyfy.pickyfy.web.dto.request;
 
 public record EmailVerificationVerifyRequest(
         String email,
