+++ src/main/java/com/pickyfy/pickyfy/web/dto/response/EmailVerificationSendResponse.java
@@ -1,4 +1,4 @@
-package com.pickyfy.pickyfy.dto.response;
+package com.pickyfy.pickyfy.web.dto.response;
 
 public record EmailVerificationSendResponse(
         String email,
